add BaseTextField class to reuse TextField
